Add TransitionDismisser.
Summary:
The transition dismisser's role is to provide a registry for gesture recognizers that can cause a view controller to be dismissed.

The API is a static method on the director class. This is because transition directors are designed to live only for the duration of their transition. Gesture recognizers need a target to send events to, so without an instance of a director to speak with (the transition hasn't started yet), we provide a TransitionDismisser instance instead.

This lifecycle guarantee encourages a clean separation of concerns between view controllers and their directors.

Reviewers: O4 Material Motion Apple platform reviewers, O2 Material Motion, markwei

Reviewed By: O4 Material Motion Apple platform reviewers, O2 Material Motion, markwei

Subscribers: markwei

Tags: #material_motion

Differential Revision: http://codereview.cc/D1971